added import and export data services and env veriables to the client side
