+++ README.md
@@ -1,1 +1,7 @@
 # GistMD-proj
+
+In this project I created an app for doctors as end user clients, where they can track their patients and add new patients to the system. 
+
+## important note
+
+I added the .env files to this project so it is available online for this project. In an actual app going for production, the .env files would have been added to the gitignore file, so it is not available online.
