+++ client/.env
@@ -1,0 +1,3 @@
+
+
+REACT_APP_PATIENT_BACKEND_ENDPOINT=http://localhost:3000/patient
